requirements.txt
opencv-python
sentence-transformers
torch
torchvision
psycopg2-binary
numpy
fastapi 
uvicorn
python-multipart